moved home and signing to pages folder
