+++ app/src/App.js
@@ -3,7 +3,7 @@ import './App.css';
 import SignIn from './components/SignIn.js';
 // import firebaseConfig from './utils/firebase/config.js'
 import firebase from 'firebase'
-import Home from './components/Home.js';
+import Home from './components/Pages/Home.js';
 
 class App extends React.Component {
   constructor(props) {
